Reworked how the blog indexes are built.
We need to let the blog article be in charge of how they are displayed. Now
summaries are build with a helper *after* the articles are compiled. This
simplified many things and allows us to show media on the index pages.

By default, the entire article will be shown. If the author wishes to show only
part, insert the HTML snippet <!--MORE-->, and everything before that will be
shown as the article summary.