Refactored data sources and identifiers involving people.
Fixes #85 and parts of #79.

Separates out the concept of a company member and any other person.
Assigns a local identifier if a web_id is not provided, so every person
  has a unique and stable identifier.
Updates relevant filters for greater clarity and to these identifiers.
Populates web_id during preprocessing as an item attribute rather than
  providing a function to dynamically assemble during buildtime.
Provides a new data source for people.