Overhaul tagging helpers.
Use the 'tag' items directly for the tag pages.
Remove the 'tag-page' kind.
Add tests for tags.
Rename the tag item's +tag+ attribute to +name+.
Remove the +semantic+ attribute and just query for the existance of needed
fields.
Use the path_to helper to get the path to the tag pages. Include LinkTo.