Added acronym for RIPESS and adjusted SSE definition based on RIPESS's
The term “Social Solidarity Economy”  started to be used in the late
90s.  The first meeting of what would thereafter become the RIPESS
network, took place in Lima, Peru on July 4th, 1997 and the participants
from more than 30 countries agreed that there needed to be a strong
integration between the more traditional social economy structures
(collective enterprises – a sector of the solidarity economy) and the
more holistic and alternative approaches of solidarity economy practices
and communities. In fact, while in most francophone and hispanophone
countries the expression used is “Social AND Solidarity Economy”, when
the RIPESS network was formally announced in December 2002, it chose to
eliminate the “AND” in its official name, in order to stress solidarity
economy’s aim of transformative system change, which includes going
beyond the social economy. Yet many networks continue to use the term
Solidarity Economy and institutions usually refer to SSE as Social and
Solidarity Economy.

@see http://www.ripess.org/what-is-sse/what-is-social-solidarity-economy/